backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
sqlalchemy
psycopg2-binary
google-auth
requests
python-jose[cryptography]
python-dotenv
